Adding first dockerfiles, making ports.go more dynamic
Removed all localhost from ports.go, since we cannot use them inside
the containers. The ports and URLs now set via the
environment variables, which can be found in docker-compose.yml.

The first version of Dockerfiles are written for a Context-box, which
can be found inside the Context-box directory.

Small changes inside server.go in Context-box directory, related to new
way of initilizing URLs and ports.