Implement the default values for docker-compose.yml
Default values are set in .env file and can be overridden
with flag --env-file. The "source of truth" list is implemented
in the docker-compose.yml.

Hostnames and ports are not concatenated in one variable. This
is because of the ports section in docker-compose.yml. The
ports can not be hardcoded there, therefore the <SERVICE>_PORT
variables were introduced in .env file. To keep the .env file
simple, and to allow users to change the port with only one variable,
the concatenation was not implemented.