Add new Dockerfiles and .dockerignore
Each service now has its own Dockerfile and .dockerignore.
Some of the Dockerfiles are not tested since code can not
be compiled at this stage of development. Thus, added Dockerfiles
are provisional, and they might change in the future.

The docker-compose.yml has now all services, although some of them
are commented, because of the unfinished development in
their source files.