Add new naming convention for generated images
The new naming convention will be used to push images
to a private image registry hosted on a GCP. The tag used
is environment variable from Github actions and will be
automatically incremented after each run.

The comments at the bottom of the `main.yml` are just a
draft and will change in the future.