Feature/GitHub actions (#5)
* Add main.yml file for Github actions

Created new directories, .github and workflows, used by the
Github actions. Created main.yml file, also used by Github
actions. The main.yml is a test file and will likely change
in the future.

* Fix issue with `main.yml` file

In the `main.yml` file, the step, where we were creating
images and containers via `docker-compose up` needs to be
carried out in the background, otherwise the Github Action
can not continue. Added `-d` flag sorted this issue.

Small changes in a names of steps, and comments.

* Add a small delay between `docker-compose up` and `docker ps`

Adding a small delay (5 seconds) between these steps will assure
that all containers are already up and running.

* Add small changes to `main.yml` file

Adding `docker-compose down` to stop the containers before the
Github actions clean up. Small changes in comments.

* Add new version of Github actions

The new version of `main.yml` now does NOT use `docker-compose`
because of an easier migration to other platforms similiar to
Docker in the future.

* Small change to `main.yml` syntax

* Add missing space in

* Add new version of `main.yml` for Github actions

Adding a new way of loading `.env` file. The `main.yml`
is still in a testing and will likely change in the future.

* Small change in `main.yml` for Github actions

* Change equals sign to colon in `main.yml`

* Add missing periods to `docker build` command

Adding a missing periods in `main.yml` for `docker build` command.
Small fixes in syntax of the file `main.yml`.

* Add `docker logs` to `main.yml` for Github actions

Adding a `docker logs` to `main.yml` for debugging. Small
changes with the syntax.

* Add `--name` for `docker run` in `main.yml` for Github actions

* Fix a typo in `main.yml`

* Small change in `main.yml` for github actions

* Create a network for docker containers, in order to connect them

Adding `docker network create` in `main.yml`, so containers can
connect to other containers via names.

* Add new naming convention for generated images

The new naming convention will be used to push images
to a private image registry hosted on a GCP. The tag used
is environment variable from Github actions and will be
automatically incremented after each run.

The comments at the bottom of the `main.yml` are just a
draft and will change in the future.

* Add version 1 for psuhing images to GCR

* Reorganise file structure for `main.yml`

* Add second version of Github actions `main.yml`

Adding a second version for Github actions. The second version
now includes automated pushing of the build images to a private
repository hosted on the GCP. The `docker-compose` is not being
used for various reasons.

* Small changes in the `main.yml`

Small changes in comments/names in `main.yml`. The
`docker-compose.yml` is no longer being used, therefore deleted.

* Test the new, optimized version of `main.yml`

Adding a for loops to check if files had changed, if yes,
save it to the environment variable and use it to apply
`docker build` (later `docker push`).

* Add small changes to `main.yml`

Fixing the syntax of the `main.yml`. The environment variable
now replaced by  ${{ steps.array.outputs.ARRAY }}.

* Small change in a syntax of the bash script in `main.yml`

* Test the github action

Add small comment in Dockerfile for Context-box, to see if new
image will be generated.

* Test the github action with a new `git diff`

The last test failed because `git diff --name-only` has not shown
any differences, thus no new changes were detected. Changed `git diff`
should resolve this issue.

* Test the github action with modified `git diff`

* Removing `origin/master` from `main.yml`

* Test the github action with modified `git diff`

Now using (`origin/master HEAD`)

* Test the github action with modified `git diff`

Github actions/checkout is not fetching every branch, need to
do it explicitly.

* Add `origin/` in `git diff`

* Fix typo:  ->

* Add log messages to `main.yml`

* Small change in a syntax in main.yml

* Add optimized version of Github actions

The new github action now builds/pushes new images only
if their code/Dockerfile/etc were changed. This avoids many
duplicates in image registry and speeds up the whole
process.

* Adding a check for empty array in main.yml

* Fix the conditon for empty array checking

* Update `.gitignore`

The update is specifying the path to terraform files to the ones,
that are generated by a Terraformer.

* Add white space to Dockerfile for testing

The white space is simulating a change in a files of service and
should trigger the build/push process.

* Update the `main.yml`

The images are now tagged as `latest` as well as by git commit SHA.
This will help us to test the new images.

* Test the github action

The `docker build` in `main.yml` is throwing an error, based on
number of arguments given.

* Add the copying of the exported output in `main.yml`

The exported value in `main.yml` was copied as a string literal,
not the array. The copying into the array should sort this problem.

* Changing `$GITHUB_SHA` to environment variable

The bash script is possibly confused by some characters in `-t` flag.
Changing it to environment variable should resolve this issue.

* Changing `GITHUB_SHA` to `github.sha`

* Change `GITHUB_SHA` to `GITHUB_RUN_NUMBER`

Testing the github action. Previously worked with run number, need
to verify if it still works.

* Fix typo in `main.yml`

* Add the missing `$` sign at the start of the variable

* Test the implementation of tag `latest`

* Uncommment the `docker tag` for the tag `latest` in a bashscript

* Remove old way of looping through the array in a last step

Since the output of a `array.output.ARRAY` is a string literal,
it needs to be copied in an array and used in a loop.

* Add terraform files used to initialize GCR

Adding the terraform files that will initilize bucket in a GCP project
used for storing images. The Terraform will NOT destroy the bucket
if `terraform destroy` is called.

* Fix typo in the name of the key

* Remove the `feature/GithubActions` from CI pipeline

* Add Terraformer and Builder to CI pipeline

* Remove `.tfstate` files, update the `main.yml`

* Add conditions to some of the steps in CI pipeline

The condition will assure that the steps will carry out,
even if some steps before will fail. Some modules from
platform are not ready to be compiled yet, therefore the fail
is expected.

* Add missing conditions to steps

* Add newline at the end of the `.env` file

The bash script in `main.yml` did not read the last line. The
newline at the end sorted this issue.

Co-authored-by: Miroslav Repka <miroslav.repka@berops.com>