Add specification of the working directory (#18)
The steps always start in a root directory of the project. Every
step have to specify if the working directory should be different
from the default one. This caused an error in the deployment step
of the workflow. Setting the `working-directory` at the start of the
step should solve this issue.

Co-authored-by: Miroslav Repka <miroslav.repka@berops.com>