CI/CD: stop using short-lived kubeconfig tokens
Previously, we used the get-gke-credentials with the default setting,
which generates the kubeconfig with a short-lived token.
Sometimes, this token isn't long-lived enough for the entire E2E pipeline.
This is why I'm swapping it for the GCP auth plugin within kubectl,
which renews itself whenever it needs to.

See https://github.com/google-github-actions/get-gke-credentials
use_auth_provider - (Optional) If true, use the Google Cloud auth plugin in kubectl instead of a short-lived access token. The default value is false.

Furthermore, since this method expects authentication to gcloud,
I'm changing the step order to first login to gcloud, and then configure kubectl.