Add the basis for the K8s layer of "Frontend"
This PR adds a K8s deployment of Frontend, but for now with just a placeholder "nginx image.
The real value here is the k8s-sidecar.
I've configured the K8s sidecar to watch for secrets with label "claudie.io/input-manifest".
Such secrets will automagically appear in the frontend container at path /input-manifests.
To demonstrate, I'm adding an example inputmanifest secret, which we can maintain
from now for as a showcase.