Fix: frontend crashes after gRPC fails to establish connection (#266)
* Fix: frontend crashes after gRPC fails to establish connection

The frontend service is crashing after 1st unsuccessful connection attempt.
The connection mechanism was adjusted with a retry policy that retries to connect
to the requested service within a graceful period of time. A retry policy was also
added to the RPC calls that will be made with that established connection.

These changes were made to make the frontend service more robust with connecting
to services. The deadlines for stopping the retry mechanism were chose as follows:
5 minutes for establishing the initial connection and 1 minute for the RPC calls that
are made with the connection.

* Auto commit - update kustomization.yaml

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>