Feat: rewrite frontend workflow
The frontend service works in a sequential way and checks
for changes in the manifest directory once every 5 minutes.
The workflow was rewritten such that it no longer has a timer
but instead it depends on the k8s-sidecar to notify the service
via a HTTP request. Sequential execution was also removed in favor
of concurrent execution.