Feat/Manager (#1498)
Closes https://github.com/berops/claudie/issues/1481
https://github.com/berops/claudie/issues/1392
https://github.com/berops/claudie/issues/1340
https://github.com/berops/claudie/issues/1318
https://github.com/berops/claudie/issues/1247
https://github.com/berops/claudie/issues/1065
https://github.com/berops/claudie/issues/1500
https://github.com/berops/claudie/issues/1502


This PR introduces the Manager service, which is a combination of the
context-box and scheduler service.

Manger is the brain and main entry point for claudie.
To build clusters users/services submit their configs to the manager
service. The manager creates the desired state and schedules a number of
jobs to be executed in order to achieve the desired state based on the
current state. The jobs are then picked up by the builder service.

With this new service the schema in the database has also changed to
reflect the same datastructures that we used when building the clusters.
That is the schema (on a high level) for a configuration is as follows:

```
type Config struct {
	Version  uint64                   `bson:"version"`
	Name     string                   `bson:"name"`
	K8SCtx   KubernetesContext        `bson:"kubernetesContext"`
	Manifest Manifest                 `bson:"manifest"`
	Clusters map[string]*ClusterState `bson:"clusters"`
}

type ClusterState struct {
	Current Clusters `bson:"current"`
	Desired Clusters `bson:"desired"`
	Events  Events   `bson:"events"`
	State   Workflow `bson:"state"`
}

type Clusters struct {
	K8s           []byte `bson:"k8s"`
	LoadBalancers []byte `bson:"loadBalancers"`
}
```

Since we are working with k8s clusters, and we allow multiple
loadbalancers to be attached to a single k8s cluster, the schema in the
database will now reflect this, by grouping a k8s cluster with relevant
lb clusters, we can simplify the logic in the application code, as we no
longer need to parse and match them repeatedly when building.

We also attach a state and events to this single entity. The state
simply describes where the k8s clusters and their lbs are in the
workflow, for example in the terraformer building infrastructure. The
events are tasks created by the manager that describe how to get from
the current state to the desired state. The builder service will
gradually pull events until the current state reflects the desired
state.

Additionally, a new field has been added, `Version`. This is needed to
implement optimistic locking. It is rarely the case that we really need
to lock access to a resource, since the flow of claudie has relatively
low contention at the document level in MongoDB, so with optimistic
locking we can get rid of locks in the application code and simply rely
on the versioning of the document. It will protect against concurrent
writes by only allowing a single change to be accepted where the other
changes have to repeat the read/update/write cycle, which as mentioned
does not happen that often with the way claudie works.

These changes aim to resolve a number of issues we had with claudie that
I did not encounter during testing. The changes will continue to be
monitored in the near future, and if any of the issues arise, they will
be reopened.

In addition, these changes implemented are also a prerequisite for the
change for rolling updates as described in the
https://github.com/berops/claudie/issues/1441 issue. We have agreed that
the rolling update will be done one nodepool at a time. This was not
possible with the way claudie worked (and works) until the changes in
this PR. Claudie always works on the whole config and all the clusters
in it, so gradually updating one nodepool at a time would not have been
straightforward and would have caused more technical debt in the future
than with the changes in this PR.

With the changes in this PR, rolling update of all node pools within a
single cluster should be relatively easy to implement by simply spawning
a task for each node pool. If an error occurs, this will also prevent
updates to the rest of the nodepools, allowing the cluster to function.

---------

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>
Co-authored-by: Jakub Hlavačka <jakubhlavacka123@gmail.com>