add GCP metadata server IP and hostnames to no proxy envs (#1517)
closes https://github.com/berops/claudie/issues/1514

This supports building K8s clusters with GCP VMs using an HTTP proxy.

When using an HTTP Proxy this
`metadata,metadata.google.internal,169.254.169.254,metadata.google.internal.`
has to be in the no proxy envs. Otherwise Claudie fails to build the k8s
cluster.

This PR also contains an update of `.env` file, cause it should run the
e2e tests with proxy mode turned on and using an HTTP proxy that
supports access to a limited set of domains. Once the e2e tests finish
successfully the changes in the `.env` file will be reverted.

---------

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>