Feat/templates rolling update (#1525)
Closes https://github.com/berops/claudie/issues/1441
Closes https://github.com/berops/claudie/issues/1518

This PR includes changes made from PR
https://github.com/berops/claudie/pull/1520 due to dependencies.
Closes https://github.com/berops/claudie/issues/1519
Closes https://github.com/berops/claudie/issues/1515

The proxy settings have been adjusted to `default` which will be turned
on when a cluster uses hetzner nodepools.
The proxy will always be turned on in the e2e pipeline.

This PR introduces rolling updates for nodepools for k8s clusters and
load balancer clusters as described in issue
https://github.com/berops/claudie/issues/1441.

Changes to the templates configurations in
https://github.com/berops/claudie-config have been made to prepare this
PR, as the previous templates would fail due to naming collisions and
validation on various cloud providers due to naming length of various
resources.

This PR complements the previous
https://github.com/berops/claudie/pull/1498 and
https://github.com/berops/claudie/pull/1460

---------

Co-authored-by: Jakub Hlavacka <jakub.hlavacka@berops.com>
Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>