Feat/init-process (#1527)
When executing ansible paybooks multiple child processes are spawned as
described in issue https://github.com/berops/claudie/issues/1317. The
problem is that these processes are not correctly cleaned up after the
playbook stops executing inside a docker environment. It unclear why the
subprocesses are not correctly terminated, to address this issue the
init system was added to the docker container for ansible.

The init process adopts, these "oprhaned children" left by ansible and
cleans them up so that we do not eventually run out of resources on the
docker image.

closes https://github.com/berops/claudie/issues/1317

---------

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>