fix: malformed terraform state (#1566)
closes https://github.com/berops/claudie/issues/1483
closes https://github.com/berops/claudie/issues/1275

This PR introduces changes as to how errors in building infrastructure
is handled in terraformer.

There are essentially 3 cases that needs to be taken care of.

1. K8s clusters fails to build due to invalid nodepool definitions.
2. K8s cluster succeeded, but at least 1 loadbalancer cluster have not
succeeded.
3. Both k8s and lb clusters succeeded but Dns failed for at least 1
loadbalancer

This PR fixes these cases as follows:

If there is an error in building the nodepools of the desired state
after an unsucessful terraform apply the state file is dumped and any
partially build state that was not part of the state file before the
terraform apply is destroyed, essentially performing a infrastructure
rollback to the last working state. After an error occurs during
terraform apply the current state is left as is and is not updated, thus
avoiding conflicts as described in issue
https://github.com/berops/claudie/issues/1483

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>