update kubeone version & adjust ansible-playbook (#1599)
adds `backrefs` to the update proxy envs ansible playbook. Without this
the environment variables would be appended to the end of the file each
time and after 5 runes the entire /etc/environment file would be
cluttered with these variables. With `backrefs` they will be overwritten
every time.

Closes https://github.com/berops/claudie/issues/1598
Closes https://github.com/berops/claudie/issues/1365

Based on the Kubeone 1.9 release docs, there does seem not to be any
backwards incompatibility issues

---------

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>