Fix/unreachable node (#1658)
This PR aims at resolving issues as described in
https://github.com/berops/claudie/issues/1631
Closes https://github.com/berops/claudie/issues/1512 by Increasing the
memory limit for builder to twice the amount.

If any node becomes unreachable claudie will report this error until it
is resolved.

If any unreachable nodes occurs for a loadbalancer claudie will prompt
be user to:
 - either deleted the nodepool from the loadbalancer or
 - delete the whole loadbalancer.
 
If any unreachable nodes occurs for a k8s cluster, claudie will prompt
the user to:
 - either delete the nodepool from the k8s cluster or
 - delete the node from the k8s cluster via `kubectl`

---------

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>