Feat/lb alternative names (#1693)
Closes https://github.com/berops/claudie/issues/1662


Adds alternative names support to claudie. This PR only implements the
support at the input manifest level and the terraform service.

The templates that claudie uses will have to be adjusted individually.

The templates can use the following pattern to check if the alternative
names extension is used or not

```
	// {{- if hasExtension .Data "AlternativeNamesExtension"}}
	// ... access .AlternativeNamesExtension
	// {{- end }}
```

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>