fix: deletion of nodes (#1919)
Fixes deletion of nodes in newer version of kubectl (>= 1.33)

the `export ETCDCTL_API=3` is no longer needed as the lowest supported
kuberenetes version (1.32.x) ships with `etcd version = 3.5.x`

and the env has been made as a default since version
[`3.4.x`](https://github.com/etcd-io/etcd/blob/cf5a57119f24a7f3f9eb9fa9323055231c9ebd45/CHANGELOG/CHANGELOG-3.4.md?plain=1#L947)

Also changed the dumping of the etcd members as a JSON, instead of
parsing raw `Stdout` output

The issue with the approach in master was that etcd started to print
warnings which would be parsed instead of the `Stdout` output, leading
to panics in the kuber service.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Bug Fixes**
* Adjusted ETCD environment handling to reduce inconsistent etcdctl
behavior during cluster node removal.

* **Refactor**
* Improved ETCD member discovery and removal flow with structured member
handling and clearer warnings when nodes aren’t found.

* **Chores**
  * Updated many image tags.
  * Bumped bundled kubectl installer and streamlined install steps.
  * Added Docker Buildx setup to CI.

* **Configuration**
* Enabled installation proxy and upgraded numerous test cluster node
types.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: CI/CD pipeline <CI/CD-pipeline@users.noreply.github.com>
Co-authored-by: Jakub Hlavacka <jakub.hlavacka@berops.com>